Merge changes from testing into development branch hotfix/mixer_drag

+++ zynlibs/zynseq/sequence.h
@@ -61,9 +61,9 @@ class Sequence
         bool removeTrack(size_t track);
 
         /** @brief  Get quantity of tracks in sequence
-        *   @retval size_t  Quantity of tracks
+        *   @retval uint32_t Quantity of tracks
         */
-        size_t getTracks();
+        uint32_t getTracks();
 
         /** @brief  Clear all tracks from sequence
         */
@@ -113,7 +113,7 @@ class Sequence
         *   @param  bSync True to indicate sync pulse, e.g. to sync tracks
         *   @param  dSamplesPerClock Samples per clock
         *   @retval uint8_t Bitwise flag of what clock triggers [1=track step | 2=change of state]
-        *     @note    Sequences are clocked syncronously but not locked to absolute time so depend on start time for absolute timing
+        *   @note   Sequences are clocked syncronously but not locked to absolute time so depend on start time for absolute timing
         *   @note   Will clock each track
         */
         uint8_t clock(uint32_t nTime, bool bSync, double dSamplesPerClock);
@@ -184,4 +184,3 @@ class Sequence
         bool m_bEmpty = true; // True if all patterns are emtpy (no events)
         std::string m_sName; // Sequence name
 };
-
